C-efficiency

Seeks to minimize the variance of the ratio of two parameters, for example, willingness-to-pay.

c_efficiency(p, x, didx, all = FALSE)

Arguments

p
A vector of parameters

x
A variance-covariance matrix

didx
An integer indicating the position of the denominator in `p`

all
If `all = TRUE` return a vector with variance for each ratio, else return a single number with sum of the K-1 ratios. The default is FALSE. If used for optimization, then a single number is used, but the variance for each parameter is always printed regardless of efficiency criteria used in optimization as long as the denominator is defined.

Value

A vector giving the variance of the ratio for each K-1 parameter or a single number with the sum of the variances used for optimization

References

Kessels, R., Goos, P. and Vandebroek, M., 2006, A comparison of criteria to design efficient choice experiments, Journal of Marketing Research, 43(3):409-419 Scarpa and Rose, 2008, Designs efficiency for non-market valuation with choice modelling: How to measure it, what to report and why, Australian Journal of Agricultural and Resource Economics, 52(3):253-282